split wizard from remix UI
