+++ src/lib/remix/components/Setup.svelte
@@ -3,7 +3,7 @@
   import type { AuthenticationResponse } from "$lib/models/auth";
   import type { APIResponse } from "$lib/models/ui";
   import { logError, logSuccess } from "$lib/remix/logger";
-  import { clearErrorBanner, globalState } from "$lib/state/state.svelte";
+  import { clearErrorBanner, globalState } from "$lib/remix/state/state.svelte";
   import Button from "./shared/Button.svelte";
 
   type Props = {
