+++ src/lib/wizard/index.ts
@@ -1,0 +1,4 @@
+export const initWizardPlugin = () => {
+  // when users configure a contract, the plugin gets the results.
+  // listenToContracts();
+}
