+++ src/routes/wizard.svelte
@@ -1,0 +1,4 @@
+<script lang="ts">
+</script>
+
+<p>Defender Deploy in Wizard</p>
